use smaller MAX_ARRAY_SIZE for unit test to avoid OOM error in builds
