+++ core/src/test/java/com/backblaze/b2/json/B2JsonTest.java
@@ -2793,34 +2793,6 @@ public class B2JsonTest extends B2BaseTest {
         assertEquals("b2", obj3.field77);
     }
 
-    /**
-     * Cannot directly test the case with MAX_ARRAY_SIZE being (Integer.MAX_VALUE - 8)
-     * since the real OutOfMemory error would occur without triggering the intended
-     * IOException ("Requested array size exceeds maximum limit"). This happens even
-     * when heap size is increased to 8GB. This happens probably because the JVM heap
-     * is probably heavily allocated before the OutputStream array size is anywhere
-     * close to half of MAX_ARRAY_SIZE.
-     *
-     * Instead, create a subclass of B2JsonByteArrayOutputStream with lowered
-     * MAX_ARRAY_SIZE (1000) for testing purpose: IOException will then be thrown
-     */
-    private static class B2JsonByteArrayOutputStreamForTest extends B2JsonByteArrayOutputStream {
-        private static final int MAX_ARRAY_SIZE = 1000;
-
-        @Override
-        protected void grow(int minCapacity) throws IOException {
-            // overflow-conscious code
-            int oldCapacity = buf.length;
-            int newCapacity = oldCapacity << 1;
-            if (newCapacity - minCapacity < 0)
-                newCapacity = minCapacity;
-            if (newCapacity - MAX_ARRAY_SIZE > 0) {
-                throw new IOException("Requested array size exceeds maximum limit");
-            }
-            buf = Arrays.copyOf(buf, newCapacity);
-        }
-    }
-
     /* convenience object for testing IOException("Requested array size exceeds maximum limit") */
     private static class ObjectWithSomeName {
        @B2Json.required
@@ -2860,7 +2832,7 @@ public class B2JsonTest extends B2BaseTest {
         }
         // the total size of object in bytes is 494; doubling will not cause overflow over 1000
         final ObjectWithSomeName expectedObjectWithSomeSmallName = new ObjectWithSomeName(smallName.toString());
-        final B2JsonByteArrayOutputStreamForTest b2JsonByteArrayOutputStreamForSmall = new B2JsonByteArrayOutputStreamForTest();
+        final B2JsonByteArrayOutputStreamTest.B2JsonByteArrayOutputStreamForTest b2JsonByteArrayOutputStreamForSmall = new B2JsonByteArrayOutputStreamTest.B2JsonByteArrayOutputStreamForTest();
         B2Json.get().toJson(expectedObjectWithSomeSmallName, b2JsonByteArrayOutputStreamForSmall);
 
         // rebuild the object based on the Json string
@@ -2879,7 +2851,7 @@ public class B2JsonTest extends B2BaseTest {
 
         String nullJson = null;
         try {
-            final B2JsonByteArrayOutputStreamForTest b2JsonByteArrayOutputStreamForBig = new B2JsonByteArrayOutputStreamForTest();
+            final B2JsonByteArrayOutputStreamTest.B2JsonByteArrayOutputStreamForTest b2JsonByteArrayOutputStreamForBig = new B2JsonByteArrayOutputStreamTest.B2JsonByteArrayOutputStreamForTest();
             final ObjectWithSomeName expectedObjectWithSomeBigName = new ObjectWithSomeName(bigName.toString());
 
             // total size of the object in bytes is 520; doubling capacity will cause overflow over 1000
